+++ MyVocaList/App.xaml.cs
@@ -1,21 +1,14 @@
-﻿#if ANDROID || IOS || MACCATALYST
-using HorusStudio.Maui.MaterialDesignControls;
-#endif
-
-namespace MyVocaList;
+﻿namespace MyVocaList;
 
 public partial class App : Application
 {
-	public App()
-	{
-		InitializeComponent();
-#if ANDROID || IOS || MACCATALYST
-		MaterialDesignControls.InitializeComponents();
-#endif
-	}
+    public App()
+    {
+        InitializeComponent();
+    }
 
-	protected override Window CreateWindow(IActivationState? activationState)
-	{
-		return new Window(new AppShell());
-	}
+    protected override Window CreateWindow(IActivationState? activationState)
+    {
+        return new Window(new AppShell());
+    }
 }